+++ tests/Functional/IntersperseTest.php
@@ -11,6 +11,36 @@ class IntersperseTest extends \PHPUnit\Framework\TestCase
     return [
       [\range(1, 4), 'foo', [1, 'foo', 2, 'foo', 3, 'foo', 4]],
       [['foo', 'bar'], 2, ['foo', 2, 'bar']],
+      [
+        (object) [
+          'foo' => 3.332,
+          'bar' => 'quux',
+          'baz' => null,
+        ],
+        false,
+        [
+          'foo' => 3.332,
+          '0'   => false,
+          'bar' => 'quux',
+          '1'   => false,
+          'baz' => null,
+        ],
+      ],
+      [
+        (object) [
+          false,
+          null,
+          \range(1, 3),
+        ],
+        'qux',
+        [
+          false,
+          'qux',
+          null,
+          'qux',
+          \range(1, 3),
+        ],
+      ],
     ];
   }